+++ metaflow/plugins/kfp/kfp_cli.py
@@ -1,5 +1,6 @@
-import click
 import posixpath
+
+import click
 
 from metaflow import current, decorators
 from metaflow.datastore.datastore import TransformableObject
@@ -13,9 +14,8 @@ from metaflow.package import MetaflowPackage
 from metaflow.plugins.aws.step_functions.step_functions_cli import (
     check_metadata_service_version,
 )
-from metaflow.plugins.kfp.kfp_constants import (
-    BASE_IMAGE,
-)
+from metaflow.plugins.kfp.kfp_constants import BASE_IMAGE
+from metaflow.plugins.kfp.kfp_step_init import save_step_environment_variables
 from metaflow.util import get_username
 
 
@@ -43,8 +43,6 @@ def kubeflow_pipelines(obj):
 @click.option("--task_id")
 @click.pass_obj
 def step_init(obj, run_id, step_name, passed_in_split_indexes, task_id):
-    from metaflow.plugins.kfp.kfp import save_step_environment_variables
-
     save_step_environment_variables(
         obj.datastore,
         obj.graph,
@@ -211,13 +209,9 @@ def run(
             response = flow._client.wait_for_run_completion(
                 run_pipeline_result.run_id, 500
             )
+
             if response.run.status == "Succeeded":
-                obj.echo(
-                    "Flow: {flow_name}, run link: {kfp_run_url}\n  SUCCEEDED!".format(
-                        flow_name=current.flow_name, kfp_run_url=kfp_run_url
-                    ),
-                    fg="green",
-                )
+                obj.echo("SUCCEEDED!", fg="green")
             else:
                 raise Exception(
                     "Flow: {flow_name}, run link: {kfp_run_url} FAILED!".format(
@@ -272,8 +266,8 @@ def make_flow(
 
     if package_url:
         obj.echo(
-            "Uploaded package to: {package_url}".format(package_url=package_url),
-            fg="magenta",
+            "*Uploaded package to:* {package_url}".format(package_url=package_url),
+            fg="cyan",
         )
 
     return KubeflowPipelines(
