+++ metaflow/plugins/kfp/kfp_constants.py
@@ -11,3 +11,5 @@ PASSED_IN_SPLIT_INDEXES_ENV_NAME = "PASSED_IN_SPLIT_INDEXES_ENV_NAME"
 TASK_ID_ENV_NAME = "TASK_ID_ENV_NAME"
 SPLIT_INDEX_ENV_NAME = "SPLIT_INDEX_ENV_NAME"
 INPUT_PATHS_ENV_NAME = "INPUT_PATHS_ENV_NAME"
+
+STEP_ENVIRONMENT_VARIABLES = "/tmp/step-environment-variables.sh"
