+++ metaflow/plugins/kfp/kfp_step_function.py
@@ -1,0 +1,48 @@
+def kfp_step_function(
+    datastore_root: str,
+    cmd_template: str,
+    kfp_run_id: str,
+    passed_in_split_indexes: str = '""',  # only if is_inside_foreach
+) -> list:
+    """
+    Renders and runs the cmd_template containing Metaflow step/init commands to
+    run within the container.
+    """
+    import os
+    import json
+    from subprocess import Popen
+    from metaflow.cli import logger
+
+    cmd = cmd_template.format(
+        run_id=kfp_run_id,
+        datastore_root=datastore_root,
+        passed_in_split_indexes=passed_in_split_indexes,
+    )
+
+    # TODO: Map username to KFP specific user/profile/namespace
+    # Running Metaflow
+    # KFP orchestrator -> running MF runtime (runs user code, handles state)
+    with Popen(
+        cmd,
+        shell=True,
+        universal_newlines=True,
+        executable="/bin/bash",
+        env=dict(
+            os.environ,
+            METAFLOW_USER="kfp-user",  # TODO: what should this be for a non-scheduled run?
+        ),
+    ) as process:
+        pass
+
+    if process.returncode != 0:
+        logger(f"---- Following command returned: {process.returncode}")
+        logger(cmd.replace(" && ", "\n"))
+        logger("----")
+        raise Exception("Returned: %s" % process.returncode)
+
+    # File written by kfp_decorator.py:task_finished
+    KFP_METAFLOW_FOREACH_SPLITS_PATH = "/tmp/kfp_metaflow_foreach_splits_dict.json"
+    if os.path.exists(KFP_METAFLOW_FOREACH_SPLITS_PATH):  # is a foreach step
+        with open(KFP_METAFLOW_FOREACH_SPLITS_PATH, "r") as file:
+            task_context_dict = json.load(file)
+            return task_context_dict.get("foreach_splits")
